Adding underline and bold to links
I added the underline text decoration to all the links but specified no underlining for top menu links.

I also split the line of code for the links. It used to contain code for unbothered links as well as links being hovered on. Instead, I made two lines: one for regular links, one for hover.

I added the bold font weight to the hovered links for further indication to users that they're mouse is in the right spot to click the links.