:wrench: Paypal Credit Card Field Styling and Footer Server Address updates
Added styling overrides to the Paypal Credit Card Field and the iFrame that the credit card information is embedded in to increase the height of the credit card field.  Prior the CSS would cause the height value to be computed at 4px which made the field unreadable.

Updated the footer case statement with the new server names and server addresses to help determine which front-end is currently serving content.